# Deferred data_sync and renaming to optimize drive writes