As per our discussions on PR #29 this is another method of optimizing the write performance of rdedup.

We maintain the data_sync completion before renaming the chunk into its final position but we write the data out and queue up the syncing to happen after the queue has grow to ```CHANNEL_SIZE```. Once grown we sync data in the first pass then rename in the second to avoid data sync and metadata write contentions. We are also caching what paths are queue for syncing / renaming so we don't try to write out the same data again and avoid a race condition on the renaming. Currently we cache all the paths written during the store but I can had a clear after the flush to keep it the same size as the queue at all times if memory is of concern.

To test the performance gains I tar my projects folder with 1.9 GB worth of git projects and then piped it into a fresh repo as a worst case scenario.

The result with running a 2 different drives 3 rounds each are:

**Faster HDD**
Original Method: 91 seconds
New Method: 78 seconds
Improvement: 14%

**Slower HDD**
Original Method: 192 seconds
New Method: 121 seconds
Improvement: 37%

*The times are the median of the three runs*