@dpc 2016-11-25
Oh. Great that you've thought about the race condition thing.

The results are really promising.

However the way I originally thought about it was not exactly the same. I did not want to have one flush of everything. Instead I wanted a flush of a single item, just delayed.

Pseudorust:

```
loop {
   if queue.len() > 200 { 
     let (file, tmp_path, path) = queue.pop();
     file.data_sync();
     rename(tmp_path, path);
   }
   let new_chunk = rx.recv();
   handle chunk
   queue.push(file, tmp_path, path);
}
```

The queue would be really a queue (`VecDequeu` or something), and my pseudorust misses handling of race conditions (just additional HashMap to track what's in the `queue`.

The difference here is that in current version during the `flush` the recently added files did not actually had enough time to hit the plater. I don't know if it's going to make any difference, or maybe it's actually better.

to have a real queue