@phillipCouto 2016-11-25
Hmm those are some good pointers! I am going to try your ideas and see what the tests show.