@phillipCouto 2016-11-25
Also I was thinking it may be more ideal to move the queue_path cache to the ```chunk_assembler_handle_data_with_edges``` so we don't send data through compression / encryption / write out if we knew at some earlier point that the exact same data already went through.

This is a minor improvement but savings on CPU time reduces power comsumption =)

What do you think? If memory becomes an issue we can look at replacing it with a more memory efficient option down the road.