@phillipCouto 2016-11-25
The I tried to follow your pseudo code and ended up with this:
```rust
let mut queue: VecDeque<(PathBuf, PathBuf, File)> = VecDeque::with_capacity(CHANNEL_SIZE);
        // Cache paths to make sure we don't queue up an existing block
        let mut queue_paths: HashSet<PathBuf> = HashSet::new();

        loop {
            match rx.recv().unwrap() {
                ChunkMessage::Exit => {
                    for &(ref path, ref tmp_path, ref file) in queue.iter() {
                        file.sync_data().unwrap();
                        drop(file);
                        fs::rename(&tmp_path, &path).unwrap();
                    }
                    return;
                }
                ChunkMessage::Data(data, sha256, chunk_type, _) => {
                    let path = self.chunk_path_by_digest(&sha256, chunk_type);
                    if !queue_paths.contains(&path) && !path.exists() {
                        let tmp_path = path.with_extension("tmp");
                        fs::create_dir_all(path.parent().unwrap()).unwrap();
                        let mut chunk_file = fs::File::create(&tmp_path).unwrap();
                        chunk_file.write_all(&data).unwrap();

                        // Queue chunk up for data sync and rename
                        queue_paths.insert(path.clone());
                        queue.push_back((path, tmp_path, chunk_file));
                        if queue.len() == CHANNEL_SIZE {
                            let (path, tmp_path, file) = queue.pop_front().unwrap();
                            file.sync_data().unwrap();
                            drop(file);
                            fs::rename(&tmp_path, &path).unwrap();
                        }
                    }
                }
            }
        }
```

Though it is definitely not as complex as the code I am submitting in the PR the performance after 3 runs on both hard drives is the same if not less than the current version in the master. My theory is once the queue is full it goes back to the same behaviour of reading and writing in quick succession and for the runs that were worse it is probably do to the hard drive writing one file and then jumping to another to sync and update metadata immediately.