@dpc 2016-11-25
I see. I guess you're right. Batching the syncs might be more efficient. Thanks for checking it!